pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "celery>=5.5.3",
    "channels>=4.2.2",
    "channels-redis>=4.2.1",
    "daphne>=4.0.0",
    "dj-rest-auth[with-social]>=7.0.1",
    "django-allauth>=65.8.0",
    "django-cors-headers>=4.7.0",
    "django-split-settings>=1.3.2",
    "djangorestframework>=3.16.0",
    "djangorestframework-simplejwt>=5.5.0",
    "jwt>=1.3.1",
    "kombu>=5.5.4",
    "psycopg2>=2.9.10",
    "pytest>=8.3.5",
    "pytest-asyncio>=1.0.0",
    "pytest-django>=4.11.1",
    "requests>=2.32.3",
]

[tool.pytest.ini_options]

python_files = ["test_*.py", "*_test.py", "tests.py"]
DJANGO_SETTINGS_MODULE = "core.settings.__init__"
pythonpath = ["src"]
